Move kl_divergence files one level upper
